Move alignment's parallel code to a separate file, to cut import overhead for the workers
